requirements.txt
PyYAML
django==4.0.6
django-common
django-ordered-model
pyOpenSSL
M2Crypto
Enum34
ipaddress
cffi
paramiko
cryptography
pillow
python-magic
